refactor: 도메인 모델 계층 구조 정리 및 패키지 분리
