+++ app/src/main/java/woowacourse/shopping/data/repository/CartRepository.kt
@@ -1,7 +1,7 @@
 package woowacourse.shopping.data.repository
 
-import woowacourse.shopping.model.Cart
-import woowacourse.shopping.model.CartItem
+import woowacourse.shopping.model.cart.Cart
+import woowacourse.shopping.model.cart.CartItem
 import woowacourse.shopping.model.Page
 
 interface CartRepository {
