+++ app/src/main/java/woowacourse/shopping/data/repository/PaymentRepository.kt
@@ -1,6 +1,6 @@
 package woowacourse.shopping.data.repository
 
-import woowacourse.shopping.model.Coupon
+import woowacourse.shopping.model.coupon.Coupon
 
 interface PaymentRepository {
     suspend fun getCoupons(): List<Coupon>
