+++ app/src/main/java/woowacourse/shopping/data/repository/RecentProductRepository.kt
@@ -1,8 +1,8 @@
 package woowacourse.shopping.data.repository
 
 import kotlinx.coroutines.flow.Flow
-import woowacourse.shopping.model.Product
-import woowacourse.shopping.model.Products
+import woowacourse.shopping.model.product.Product
+import woowacourse.shopping.model.product.Products
 
 interface RecentProductRepository {
     suspend fun getRecentProducts(): Products
