+++ app/src/main/java/woowacourse/shopping/data/repository/network/RetrofitCartRepository.kt
@@ -5,8 +5,8 @@ import woowacourse.shopping.data.remote.dto.Quantity
 import woowacourse.shopping.data.remote.dto.toDomain
 import woowacourse.shopping.data.remote.service.CartService
 import woowacourse.shopping.data.repository.CartRepository
-import woowacourse.shopping.model.Cart
-import woowacourse.shopping.model.CartItem
+import woowacourse.shopping.model.cart.Cart
+import woowacourse.shopping.model.cart.CartItem
 import woowacourse.shopping.model.Page
 
 class RetrofitCartRepository(
