+++ app/src/main/java/woowacourse/shopping/data/repository/network/RetrofitPaymentRepository.kt
@@ -3,7 +3,7 @@ package woowacourse.shopping.data.repository.network
 import woowacourse.shopping.data.remote.dto.toDomain
 import woowacourse.shopping.data.remote.service.PaymentService
 import woowacourse.shopping.data.repository.PaymentRepository
-import woowacourse.shopping.model.Coupon
+import woowacourse.shopping.model.coupon.Coupon
 
 class RetrofitPaymentRepository(
     private val service: PaymentService
