+++ app/src/main/java/woowacourse/shopping/model/User.kt
@@ -1,5 +1,7 @@
 package woowacourse.shopping.model
 
+import woowacourse.shopping.model.cart.Cart
+
 data class User(
     val email: String,
     val password: String,
