+++ app/src/main/java/woowacourse/shopping/model/coupon/CouponContext.kt
@@ -1,5 +1,7 @@
-package woowacourse.shopping.model
+package woowacourse.shopping.model.coupon
 
+import woowacourse.shopping.model.Money
+import woowacourse.shopping.model.cart.CartItem
 import java.time.Clock
 
 data class CouponContext(
