+++ app/src/main/java/woowacourse/shopping/ui/cart/component/CartRecommendationBody.kt
@@ -17,7 +17,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import woowacourse.shopping.model.Money
-import woowacourse.shopping.model.Product
+import woowacourse.shopping.model.product.Product
 import woowacourse.shopping.ui.common.model.ProductUiModel
 import woowacourse.shopping.ui.common.theme.Gray5
 import woowacourse.shopping.ui.shopping.component.ProductUnit
