+++ app/src/main/java/woowacourse/shopping/ui/common/model/ProductUiModel.kt
@@ -1,6 +1,6 @@
 package woowacourse.shopping.ui.common.model
 
-import woowacourse.shopping.model.Product
+import woowacourse.shopping.model.product.Product
 
 data class ProductUiModel(
     val product: Product,
