+++ app/src/main/java/woowacourse/shopping/ui/productdetail/component/LastViewedProductBanner.kt
@@ -16,7 +16,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import woowacourse.shopping.model.Money
-import woowacourse.shopping.model.Product
+import woowacourse.shopping.model.product.Product
 
 @Composable
 fun LastViewedProductBanner(
