+++ app/src/test/java/woowacourse/shopping/recommender/ProductRecommenderTest.kt
@@ -5,9 +5,9 @@ package woowacourse.shopping.recommender
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.Test
-import woowacourse.shopping.model.CartItem
+import woowacourse.shopping.model.cart.CartItem
 import woowacourse.shopping.model.Money
-import woowacourse.shopping.model.Product
+import woowacourse.shopping.model.product.Product
 
 class ProductRecommenderTest {
     private val electronics = "electronics"
